Add get product api

+++ src/main/java/org/example/shoppingapp/dao/AbstractHibernateDao.java
@@ -26,7 +26,7 @@ public abstract class AbstractHibernateDao<T> {
         return session.createQuery(criteria).getResultList();
     }
 
-    public T findById(int id) {
+    public T findById(long id) {
         return getCurrentSession().get(classType, id);
     }
